+++ convex-core/src/main/java/convex/core/cvm/AccountStatus.java
@@ -42,7 +42,7 @@ public class AccountStatus extends ARecordGeneric {
 	private final Address parent;
 	
 	// Sequence of fields in account. 
-	private static final Keyword[] ACCOUNT_KEYS = new Keyword[] { Keywords.SEQUENCE, Keywords.KEY, 
+	public static final Keyword[] ACCOUNT_KEYS = new Keyword[] { Keywords.SEQUENCE, Keywords.KEY, 
 			Keywords.BALANCE,Keywords.ALLOWANCE,
 			Keywords.HOLDINGS, Keywords.CONTROLLER,
 			Keywords.ENVIRONMENT,Keywords.METADATA,
@@ -367,7 +367,7 @@ public class AccountStatus extends ARecordGeneric {
 	}
 	
 	/**
-	 * Gets the Environment for this account. Defaults to the an empty map if no Environment has been created.
+	 * Gets the Environment for this account.
 	 * @return Environment map for this Account
 	 */
 	@SuppressWarnings("unchecked")