+++ convex-core/src/test/java/convex/core/lang/RTTest.java
@@ -3,17 +3,23 @@ package convex.core.lang;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertSame;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
 
 import convex.core.cvm.Address;
 import convex.core.cvm.Keywords;
+import convex.core.cvm.Symbols;
 import convex.core.data.AList;
 import convex.core.data.AVector;
 import convex.core.data.Keyword;
 import convex.core.data.Lists;
+import convex.core.data.Sets;
 import convex.core.data.Strings;
 import convex.core.data.Symbol;
 import convex.core.data.Vectors;
@@ -29,6 +35,7 @@ import static convex.test.Assertions.*;
  * these are useful for testing utility functions, edge cases and internal
  * behaviour of the RT class itself.
  */
+@TestInstance(Lifecycle.PER_CLASS)
 public class RTTest {
 
 	@Test
@@ -103,9 +110,35 @@ public class RTTest {
 	}
 	
 	@Test
+	public void testJVMCasts() {
+		assertEquals((Long)1L, RT.jvm(CVMLong.ONE));
+		assertEquals((Double)1.0, RT.jvm(CVMDouble.ONE));
+		assertEquals("foo", RT.jvm(Symbols.FOO));
+		assertEquals("foo", RT.jvm(Keywords.FOO));
+		
+		{
+			List<Object> e=RT.jvm(Vectors.empty());
+			assertTrue(e.isEmpty());
+		}
+		
+		{
+			List<Object> e=RT.jvm(Sets.empty());
+			assertTrue(e.isEmpty());
+		}
+		
+		{
+			List<Object> e=RT.jvm(Lists.empty());
+			assertTrue(e.isEmpty());
+		}
+
+
+	}
+	
+	@Test
 	public void testGetIn() {
 		assertNull(RT.getIn(null, 1));
 		assertCVMEquals(3,RT.getIn(Vectors.of(2,3), 1));
+		assertCVMEquals(3,Vectors.of(2,3).getIn(1));
 	}
 
 	@Test